LayoutTests/fast/canvas/CanvasRenderingContext2DSettings-colorSpace-enabled-expected.txt

Test that the colorSpace member of CanvasRenderingContext2DSettings is supported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSettings.colorSpace is "srgb"
PASS testSettings.colorSpace is "srgb"
PASS document.createElement("canvas").getContext("2d", { colorSpace: "foo" }) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
